// src/types/generated/ctp.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** The `BigDecimal` scalar type represents signed fractional values with arbitrary precision. */
  BigDecimal: string;
  /** [ISO 3166-1](http://en.wikipedia.org/wiki/ISO_3166-1) country code. */
  Country: string;
  /** Represents a currency. Currencies are identified by their [ISO 4217](http://www.iso.org/iso/home/standards/currency_codes.htm) currency codes. */
  Currency: string;
  /** DateTime is a scalar value that represents an ISO8601 formatted date. */
  Date: string;
  /** DateTime is a scalar value that represents an ISO8601 formatted date and time. */
  DateTime: string;
  /** Raw JSON value */
  Json: { [key: string]: unknown };
  /** A key that references a resource. */
  KeyReferenceInput: string;
  /** Locale is a scalar value represented as a string language tag. */
  Locale: string;
  /** The `Long` scalar type represents non-fractional signed whole numeric values. Long can represent values between -(2^63) and 2^63 - 1. */
  Long: number;
  /** Search filter. It is represented as a string and has th same format as in REST API: "field:filter_criteria" */
  SearchFilter: string;
  /** Search sort */
  SearchSort: string;
  /** A set. */
  Set: unknown[];
  /** Time is a scalar value that represents an ISO8601 formatted time. */
  Time: string;
  /** YearMonth is a scalar value that represents an ISO8601 formatted year and month. */
  YearMonth: string;
};
// ...
export enum TChannelRole {
  /** Role tells that this channel can be used to track inventory entries.Channels with this role can be treated as warehouses */
  InventorySupply = 'InventorySupply',
  /** Role tells that this channel can be used to track order export activities. */
  OrderExport = 'OrderExport',
  /** Role tells that this channel can be used to track order import activities. */
  OrderImport = 'OrderImport',
  /** This role can be combined with some other roles (e.g. with `InventorySupply`) to represent the fact that this particular channel is the primary/master channel among the channels of the same type. */
  Primary = 'Primary',
  /** Role tells that this channel can be used to expose products to a specific distribution channel. It can be used by the cart to select a product price. */
  ProductDistribution = 'ProductDistribution',
}
// ...
export type TCustomFieldsType = {
  __typename?: 'CustomFieldsType';
  /** This field contains non-typed data. */
  customFieldsRaw?: Maybe<Array<TRawCustomField>>;
  type?: Maybe<TTypeDefinition>;
  typeRef: TReference;
};
// ...
export type TCustomer = TReferenceExpandable &
  TVersioned & {
    __typename?: 'Customer';
    addresses: Array<TAddress>;
    authenticationMode?: Maybe<TAuthenticationMode>;
    billingAddressIds: Array<Scalars['String']>;
    billingAddresses: Array<TAddress>;
    companyName?: Maybe<Scalars['String']>;
    createdAt: Scalars['DateTime'];
    createdBy?: Maybe<TInitiator>;
    custom?: Maybe<TCustomFieldsType>;
    customerGroup?: Maybe<TCustomerGroup>;
    customerGroupRef?: Maybe<TReference>;
    customerNumber?: Maybe<Scalars['String']>;
    dateOfBirth?: Maybe<Scalars['Date']>;
    defaultBillingAddress?: Maybe<TAddress>;
    defaultBillingAddressId?: Maybe<Scalars['String']>;
    defaultShippingAddress?: Maybe<TAddress>;
    defaultShippingAddressId?: Maybe<Scalars['String']>;
    email: Scalars['String'];
    externalId?: Maybe<Scalars['String']>;
    firstName?: Maybe<Scalars['String']>;
    id: Scalars['String'];
    isEmailVerified: Scalars['Boolean'];
    key?: Maybe<Scalars['String']>;
    lastModifiedAt: Scalars['DateTime'];
    lastModifiedBy?: Maybe<TInitiator>;
    lastName?: Maybe<Scalars['String']>;
    locale?: Maybe<Scalars['Locale']>;
    middleName?: Maybe<Scalars['String']>;
    password?: Maybe<Scalars['String']>;
    salutation?: Maybe<Scalars['String']>;
    shippingAddressIds: Array<Scalars['String']>;
    shippingAddresses: Array<TAddress>;
    stores: Array<TStore>;
    storesRef: Array<TKeyReference>;
    title?: Maybe<Scalars['String']>;
    vatId?: Maybe<Scalars['String']>;
    version: Scalars['Long'];
  };
// ...
export type TInitiator = {
  __typename?: 'Initiator';
  anonymousId?: Maybe<Scalars['String']>;
  clientId?: Maybe<Scalars['String']>;
  customerRef?: Maybe<TReference>;
  externalUserId?: Maybe<Scalars['String']>;
  isPlatformClient?: Maybe<Scalars['Boolean']>;
  userRef?: Maybe<TReference>;
};
// ...
export type TReference = {
  __typename?: 'Reference';
  id: Scalars['String'];
  typeId: Scalars['String'];
};

export type TReferenceAttribute = TAttribute & {
  __typename?: 'ReferenceAttribute';
  id: Scalars['String'];
  name: Scalars['String'];
  typeId: Scalars['String'];
};

export type TReferenceAttributeDefinitionType = TAttributeDefinitionType & {
  __typename?: 'ReferenceAttributeDefinitionType';
  name: Scalars['String'];
  referenceTypeId: Scalars['String'];
};

export type TReferenceExpandable = {
  id: Scalars['String'];
};
// ...
export type TReview = TReferenceExpandable &
  TVersioned & {
    __typename?: 'Review';
    authorName?: Maybe<Scalars['String']>;
    createdAt: Scalars['DateTime'];
    createdBy?: Maybe<TInitiator>;
    custom?: Maybe<TCustomFieldsType>;
    customer?: Maybe<TCustomer>;
    customerRef?: Maybe<TReference>;
    id: Scalars['String'];
    includedInStatistics: Scalars['Boolean'];
    key?: Maybe<Scalars['String']>;
    lastModifiedAt: Scalars['DateTime'];
    lastModifiedBy?: Maybe<TInitiator>;
    locale?: Maybe<Scalars['Locale']>;
    rating?: Maybe<Scalars['String']>;
    state?: Maybe<TState>;
    stateRef?: Maybe<TReference>;
    target?: Maybe<TReviewTarget>;
    targetRef?: Maybe<TReference>;
    text?: Maybe<Scalars['String']>;
    title?: Maybe<Scalars['String']>;
    uniquenessValue?: Maybe<Scalars['String']>;
    version: Maybe<Scalars['String']>;
  };
// ...
export type TReviewTarget = {
  id: Scalars['String'];
};

export type TReviewUpdateAction = {
  setAuthorName?: InputMaybe<TSetReviewAuthorName>;
  setCustomField?: InputMaybe<TSetReviewCustomField>;
  setCustomType?: InputMaybe<TSetReviewCustomType>;
  setCustomer?: InputMaybe<TSetReviewCustomer>;
  setKey?: InputMaybe<TSetReviewKey>;
  setLocale?: InputMaybe<TSetReviewLocale>;
  setRating?: InputMaybe<TSetReviewRating>;
  setTarget?: InputMaybe<TSetReviewTarget>;
  setText?: InputMaybe<TSetReviewText>;
  setTitle?: InputMaybe<TSetReviewTitle>;
  transitionState?: InputMaybe<TTransitionReviewState>;
};
// ...
export type TSetReviewAuthorName = {
  authorName?: InputMaybe<Scalars['String']>;
};

export type TSetReviewCustomField = {
  name: Scalars['String'];
  value?: InputMaybe<Scalars['String']>;
};

export type TSetReviewCustomType = {
  fields?: InputMaybe<Array<TCustomFieldInput>>;
  type?: InputMaybe<TResourceIdentifierInput>;
  typeId?: InputMaybe<Scalars['String']>;
  typeKey?: InputMaybe<Scalars['String']>;
};

export type TSetReviewCustomer = {
  customer?: InputMaybe<TResourceIdentifierInput>;
};

export type TSetReviewKey = {
  key?: InputMaybe<Scalars['String']>;
};

export type TSetReviewLocale = {
  locale?: InputMaybe<Scalars['Locale']>;
};

export type TSetReviewRating = {
  rating?: InputMaybe<Scalars['Int']>;
};

export type TSetReviewTarget = {
  target?: InputMaybe<TTargetReferenceInput>;
};

export type TSetReviewText = {
  text?: InputMaybe<Scalars['String']>;
};

export type TSetReviewTitle = {
  title?: InputMaybe<Scalars['String']>;
};
// ...
export type TState = TReferenceExpandable &
  TVersioned & {
    __typename?: 'State';
    builtIn: Scalars['Boolean'];
    createdAt: Scalars['DateTime'];
    createdBy?: Maybe<TInitiator>;
    description?: Maybe<Scalars['String']>;
    descriptionAllLocales?: Maybe<Array<TLocalizedString>>;
    id: Scalars['String'];
    initial: Scalars['Boolean'];
    key?: Maybe<Scalars['String']>;
    lastModifiedAt: Scalars['DateTime'];
    lastModifiedBy?: Maybe<TInitiator>;
    name?: Maybe<Scalars['String']>;
    nameAllLocales?: Maybe<Array<TLocalizedString>>;
    roles: Array<TStateRole>;
    transitions?: Maybe<Array<TState>>;
    transitionsRef?: Maybe<Array<TReference>>;
    type: TStateType;
    version: Scalars['Long'];
  };
// ...
export type TTransitionReviewState = {
  force?: InputMaybe<Scalars['Boolean']>;
  state: TResourceIdentifierInput;
};
// ...
export type TVersioned = {
  createdAt: Scalars['DateTime'];
  createdBy?: Maybe<TInitiator>;
  id: Scalars['String'];
  lastModifiedAt: Scalars['DateTime'];
  lastModifiedBy?: Maybe<TInitiator>;
  version: Scalars['Long'];
};
// ...
export type TFetchReviewsQueryVariables = Exact<{
  limit: Scalars['Int'];
  offset: Scalars['Int'];
  sort?: InputMaybe<Array<Scalars['String']> | Scalars['String']>;
}>;

export type TFetchReviewsQuery = {
  __typename?: 'Query';
  reviews: {
    __typename?: 'ReviewQueryResult';
    total: number;
    count: number;
    offset: number;
    results: Array<{
      __typename?: 'Review';
      id: string;
      key: string;
      rating: string,
      roles: Array<TChannelRole>;
      nameAllLocales?: Array<{
        __typename?: 'LocalizedString';
        locale: string;
        value: string;
      }> | null;
    }>;
  };
};

export type TFetchReviewDetailsQueryVariables = Exact<{
  reviewId: Scalars['String'];
}>;

export type TFetchReviewDetailsQuery = {
  __typename?: 'Query';
  review?: TReview
};

export type TUpdateReviewDetailsMutationVariables = Exact<{
  reviewId: Scalars['String'];
  version: Scalars['Long'];
  actions: Array<TReviewUpdateAction> | TReviewUpdateAction;
}>;

export type TUpdateReviewDetailsMutation = {
  __typename?: 'Mutation';
  updateReview?: {
    __typename?: 'Review';
    id: string;
    version: number;
    key: string;
    rating: string;
    nameAllLocales?: Array<{
      __typename?: 'LocalizedString';
      locale: string;
      value: string;
    }> | null;
  } | null;
};